[BE] refactor: 엔티티 파일 폴더 위치 변경
- common없애고 techstack , category 분리
